+++ database/migrations/2025_04_18_172015_create_order_product_table.php
@@ -11,17 +11,21 @@ return new class extends Migration
      */
     public function up(): void
     {
+        if (Schema::hasTable('order_product')) {
+            return;
+        }
+
         Schema::create('order_product', function (Blueprint $table) {
             $table->id();
             $table->unsignedBigInteger('order_id')->nullable();
             $table->unsignedBigInteger('product_id')->nullable();
             $table->integer('quantity')->default(1); // optional
             $table->timestamps();
-        
+
             $table->foreign('order_id')->references('id')->on('orders')->onDelete('cascade');
             $table->foreign('product_id')->references('id')->on('products')->onDelete('set null');
         });
-        
+
     }
 
     /**
