+++ database/migrations/2025_05_03_213339_add_department_id_to_products_table.php
@@ -12,11 +12,19 @@ return new class extends Migration
     public function up(): void
     {
         Schema::table('products', function (Blueprint $table) {
-            $table->unsignedBigInteger('department_id')->nullable();
-            $table->foreign('department_id')->references('id')->on('departments')->onDelete('cascade');
-            $table->unsignedBigInteger('sub_department_id')->nullable();
-            $table->foreign('sub_department_id')->references('id')->on('sub_departments')->onDelete('cascade');
-       
+            if (! Schema::hasColumn('products', 'department_id')) {
+                $table->unsignedBigInteger('department_id')->nullable();
+            }
+            if (! Schema::hasColumn('products', 'department_id')) {
+                $table->foreign('department_id')->references('id')->on('departments')->onDelete('cascade');
+            }
+            if (! Schema::hasColumn('products', 'sub_department_id')) {
+                $table->unsignedBigInteger('sub_department_id')->nullable();
+            }
+            if (! Schema::hasColumn('products', 'sub_department_id')) {
+                $table->foreign('sub_department_id')->references('id')->on('sub_departments')->onDelete('cascade');
+            }
+
         });
     }
