+++ database/migrations/2025_05_03_221045_change_columns_in_partnerships_table.php
@@ -11,9 +11,15 @@ return new class extends Migration
      */
     public function up(): void
     {
+        if (Schema::hasColumn('partnerships', 'department_id')) {
+            return;
+        }
+
         Schema::table('partnerships', function (Blueprint $table) {
-            $table->dropForeign('partners_user_id_foreign');
-            $table->dropColumn('user_id');
+            if (Schema::hasColumn('partnerships', 'user_id')) {
+                $table->dropForeign('partners_user_id_foreign');
+                $table->dropColumn('user_id');
+            }
             $table->unsignedBigInteger('department_id')->nullable();
             $table->foreign('department_id')->references('id')->on('departments')->onDelete('cascade');
             $table->unsignedBigInteger('sub_department_id')->nullable();
@@ -22,7 +28,6 @@ return new class extends Migration
             $table->foreign('product_id')->references('id')->on('products')->onDelete('cascade');
             $table->string('status')->default('ongoing');
             $table->string('type')->nullable();
-        
         });
     }
