+++ database/migrations/2025_05_03_222736_add_partner_id_to_partnerships_table.php
@@ -12,8 +12,12 @@ return new class extends Migration
     public function up(): void
     {
         Schema::table('partnerships', function (Blueprint $table) {
-            $table->unsignedBigInteger('partner_id')->nullable();
-            $table->foreign('partner_id')->references('id')->on('partners')->onDelete('cascade');
+            if (! Schema::hasColumn('partnerships', 'partner_id')) {
+                $table->unsignedBigInteger('partner_id')->nullable();
+            }
+            if (! Schema::hasColumn('partnerships', 'partner_id')) {
+                $table->foreign('partner_id')->references('id')->on('partners')->onDelete('cascade');
+            }
 
         });
     }
